requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydub==0.25.1
google-generativeai==0.7.1
python-multipart==0.0.6